# Optimize masked Dynamic Chunk Convolution