--- a/speechbrain/lobes/models/transformer/Conformer.py
+++ b/speechbrain/lobes/models/transformer/Conformer.py
@@ -13,7 +13,6 @@ import torch.nn as nn
 import torch.nn.functional as F
 from typing import Optional, List
 import speechbrain as sb
-import math
 import warnings