## What does this PR do?

#2140 introduced a streaming Conformer-Transducer recipe. @TParcollet noted a slowdown when training with `streaming: True`. This PR reworks the convolution module code to not use lists of slices and expresses the entire chunking logic as tensor operations.

- The layernorm and bottleneck are unaffected by the time axis, and are thus applied over the input tensor directly.
- The "list of slice + stack" logic was fully replaced in favor of more clever use of padding and `Tensor.unfold`.

When testing, I found the output of the function to be virtually identical without and with this PR, and the WER% remains unchanged. Training doesn't crash. No breaking changes should occur due to the PR.

I tested performance on my 7900 XT (HIP vs CUDA might have some performance characteristics differences) in a more extreme case (chunk size of `8`, longest batch in `test-clean`, no GS, forward-only):

| Scenario | Performance |
|-|-|
| No masking | `124 ms ± 212 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)` |
| Chunk size = 8, **before** | `236 ms ± 1.55 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)` |
| Chunk size = 8, **after** | `186 ms ± 645 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)` |

The difference is fairly large here as the chunk size is rather small. I don't exactly know how this would translate in training performance and whether that fully solves the performance difference observed by @TParcollet.

<details>
  <summary><b>Before submitting</b></summary>

- [x] Did you read the [contributor guideline](https://speechbrain.readthedocs.io/en/latest/contributing.html)?
- [x] Did you make sure your **PR does only one thing**, instead of bundling different changes together?
- [x] Did you make sure to **update the documentation** with your changes? (if necessary)
- [x] Did you write any **new necessary tests**? (not for typos and docs)
- [x] Did you verify new and **existing [tests](https://github.com/speechbrain/speechbrain/tree/develop/tests) pass** locally with your changes?
- [x] Did you list all the **breaking changes** introduced by this pull request?
- [x] Does your code adhere to project-specific code style and conventions?

</details>

## PR review

<details>
  <summary>Reviewer checklist</summary>

- [x] Is this pull request ready for review? (if not, please submit in draft mode)
- [x] Check that all items from **Before submitting** are resolved
- [x] Make sure the title is self-explanatory and the description concisely explains the PR
- [x] Add labels and milestones (and optionally projects) to the PR so it can be classified
- [x] Confirm that the changes adhere to compatibility requirements (e.g., Python version, platform)
- [x] Review the self-review checklist to ensure the code is ready for review

</details>