@TParcollet 2023-12-19
Should we unit-test this as well to make sure both methods are equivalents?