@asumagic 2023-12-20
> Should we unit-test this as well to make sure both methods are equivalents?

Actually a good point, should've done so from the start. I've just implemented a test to test this at `ConformerLayer` level.