@TParcollet 2024-01-02
I am re-running the experiment now and comparing it, I'll keep you posted. The code LGTM for me.